requirements.txt
streamlit==1.31.0
ultralytics==8.1.0
opencv-python-headless==4.9.0.80
numpy==1.24.3
pillow==10.2.0
streamlit-webrtc==0.47.1
av==11.0.0
